matched count

+++ tests/communication_tests.cpp
@@ -20,7 +20,7 @@ int main(int argc, char **argv) {
   int rank;
   MPI_Comm_rank(MPI_COMM_WORLD, &rank);
 
-  vector<int> random_number_vec = generate_random_numbers(0,60000,1234,10);
+  vector<uint64_t> random_number_vec = generate_random_numbers(0,60000,1234,10);
   string output_path = "random_number_generators_"+ to_string(rank)+".txt";
   char stats[500];
   strcpy(stats, output_path.c_str());
